requirements.txt
fastapi
torch==2.1.2
uvicorn
pillow
opencv-python
numpy
catboost
scipy
psycopg2-binary
python-multipart
ultralytics==8.0.176
